Construct PoSChainConfig explicitly for testnet
