+++ common/src/chain/config/builder.rs
@@ -22,15 +22,19 @@ use crate::{
             emission_schedule, ChainConfig, ChainType, EmissionScheduleTabular,
         },
         get_initial_randomness,
-        pos::config_builder::PoSChainConfigBuilder,
+        pos::{
+            DEFAULT_BLOCK_COUNT_TO_AVERAGE, DEFAULT_MATURITY_DISTANCE, DEFAULT_TARGET_BLOCK_TIME,
+        },
         pos_initial_difficulty,
         pow::PoWChainConfigBuilder,
-        CoinUnit, ConsensusUpgrade, Destination, GenBlock, Genesis, NetUpgrades,
+        CoinUnit, ConsensusUpgrade, Destination, GenBlock, Genesis, NetUpgrades, PoSChainConfig,
         PoSConsensusVersion, PoWChainConfig, UpgradeVersion,
     },
     primitives::{
-        id::WithId, semver::SemVer, Amount, BlockDistance, BlockHeight, Id, Idable, H256,
+        id::WithId, per_thousand::PerThousand, semver::SemVer, Amount, BlockDistance, BlockHeight,
+        Id, Idable, H256,
     },
+    Uint256,
 };
 use crypto::key::hdkd::child_number::ChildNumber;
 
@@ -63,6 +67,10 @@ impl ChainType {
                 NetUpgrades::initialize(upgrades).expect("net upgrades")
             }
             ChainType::Testnet => {
+                let target_block_time = DEFAULT_TARGET_BLOCK_TIME;
+                let target_limit = (Uint256::MAX / Uint256::from_u64(target_block_time.get()))
+                    .expect("Target block time cannot be zero as per NonZeroU64");
+
                 let upgrades = vec![
                     (
                         BlockHeight::new(0),
@@ -74,9 +82,15 @@ impl ChainType {
                             initial_difficulty: Some(
                                 pos_initial_difficulty(ChainType::Testnet).into(),
                             ),
-                            config: PoSChainConfigBuilder::new(ChainType::Testnet)
-                                .consensus_version(PoSConsensusVersion::V0)
-                                .build(),
+                            config: PoSChainConfig::new(
+                                target_limit,
+                                target_block_time,
+                                DEFAULT_MATURITY_DISTANCE,
+                                DEFAULT_MATURITY_DISTANCE,
+                                DEFAULT_BLOCK_COUNT_TO_AVERAGE,
+                                PerThousand::new(1).expect("must be valid"),
+                                PoSConsensusVersion::V0,
+                            ),
                         }),
                     ),
                     (
@@ -84,11 +98,15 @@ impl ChainType {
                         BlockHeight::new(9999999999),
                         UpgradeVersion::ConsensusUpgrade(ConsensusUpgrade::PoS {
                             initial_difficulty: None,
-                            config: PoSChainConfigBuilder::new(ChainType::Testnet)
-                                .consensus_version(PoSConsensusVersion::V1)
-                                .decommission_maturity_distance(BlockDistance::new(7200))
-                                .spend_share_maturity_distance(BlockDistance::new(7200))
-                                .build(),
+                            config: PoSChainConfig::new(
+                                target_limit,
+                                target_block_time,
+                                BlockDistance::new(7200),
+                                BlockDistance::new(7200),
+                                DEFAULT_BLOCK_COUNT_TO_AVERAGE,
+                                PerThousand::new(1).expect("must be valid"),
+                                PoSConsensusVersion::V1,
+                            ),
                         }),
                     ),
                 ];
